Replace Module Source Interface by abstract class
