+++ classes/UpgradeTools/Module/Source/Provider/LocalSourceProvider.php
@@ -36,13 +36,10 @@ use ZipArchive;
  * Get the list of module zips in admin/autoupgrade/modules
  * These zips will be used to upgrade related modules instead of using distant zips on addons
  */
-class LocalSourceProvider implements ModuleSourceProviderInterface
+class LocalSourceProvider extends AbstractModuleSourceProvider
 {
     /** @var string */
     private $sourceFolder;
-
-    /** @var ModuleSource[]|null */
-    private $localModuleZips;
 
     /** @var FileConfigurationStorage */
     private $fileConfigurationStorage;
